ny filstruktur samt delat upp vue i komponenter samt delat upp flask-appen i rutter
