+++ Traffik_projekt/backend/models/supabase_client.py
