+++ Traffik_projekt/backend/routes/signup.py
@@ -1,11 +1,9 @@
-from flask import Flask, request, jsonify
-from flask_cors import CORS
+from flask import Blueprint, request, jsonify
 from models.supabase_client import supabase
 
-app = Flask(__name__)
-CORS(app)  # Tillåt frontend att anropa
+signup_blueprint = Blueprint('signup', __name__)
 
-@app.route('/api/signup', methods=['POST'])
+@signup_blueprint.route('/api/signup', methods=['POST'])
 def signup():
     data = request.get_json()
     email = data.get("email")
@@ -19,7 +17,3 @@ def signup():
         return jsonify({"message": "User created", "email": email}), 200
     except Exception as e:
         return jsonify({"error": str(e)}), 400
-
-if __name__ == "__main__":
-    print("Flask körs på http://localhost:5000")
-    app.run(debug=True)
